app/requirements.txt
fastapi
sqlmodel
uvicorn
python-dotenv
pytest
